requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
reportlab>=3.6.0
Pillow>=9.0.0
pathlib2>=2.3.0
